01/11

+++ go/status/status.go
@@ -6,9 +6,9 @@ type TCP struct {
 	DstIP   string `JSON:"dstIP"`
 	SrcMac  string `JSON:"srcMac"`
 	DstMac  string `JSON:"dstMac"`
-	SrcPort uint16 `JSON:"srcPort"`
-	DstPort uint16 `JSON:"dstPort"`
-	Timeout int64  `JSON:"timeout"`
+	SrcPort int    `JSON:"srcPort"` // uint16
+	DstPort int    `JSON:"dstPort"` // uint16
+	Timeout int    `JSON:"timeout"` // int64
 	Times   int    `JSON:"times"`
 }
